// src/types/contact.ts
export interface Contact {
  id: string;
  last_name: string | null;
  first_name: string | null;
  email: string | null;
  address: string | null;
  zipcode: string | null;
  city: string | null;
  birth_date: string | null;
  birth_department: string | null;
  phone: string | null;
  created_at: string;
}